Rename tax-report-dialog
